+++ test/source/test_ilds.cpp
@@ -37,7 +37,7 @@ TEST_CASE("VdCorput_i different base and scale") {
 }
 
 TEST_CASE("Halton_i") {
-    const std::array<size_t, 2> base = {2, 3};
+    const std::array<unsigned long, 2> base = {2, 3};
     const std::array<unsigned int, 2> scale = {11, 7};
     auto hgen = ildsgen::Halton(base, scale);
     const auto arr = hgen.pop();
@@ -50,7 +50,7 @@ TEST_CASE("Halton_i") {
 }
 
 TEST_CASE("Halton_i different bases and scales") {
-    const std::array<size_t, 2> base = {3, 5};
+    const std::array<unsigned long, 2> base = {3, 5};
     const std::array<unsigned int, 2> scale = {6, 4};
     auto hgen = ildsgen::Halton(base, scale);
     hgen.reseed(0);
