+++ test/source/test_rapidcheck.cpp
@@ -8,8 +8,8 @@
 
 TEST_CASE("Property-based test: VdCorput values in [0,1]") {
     rc::check("vdc(n, base) returns values in [0, 1]", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 101));
-        for (size_t i = 0; i < 100; ++i) {
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 101));
+        for (int i = 0; i < 100; ++i) {
             auto val = ldsgen::vdc(i, base);
             RC_ASSERT(val >= 0.0 && val < 1.0);
         }
@@ -18,8 +18,8 @@ TEST_CASE("Property-based test: VdCorput values in [0,1]") {
 
 TEST_CASE("Property-based test: VdCorput sequence is deterministic") {
     rc::check("VdCorput produces same sequence after reseed", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto seed = static_cast<size_t>(*rc::gen::inRange(0, 1000));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto seed = static_cast<unsigned long>(*rc::gen::inRange(0, 1000));
 
         ldsgen::VdCorput gen1(base);
         ldsgen::VdCorput gen2(base);
@@ -27,7 +27,7 @@ TEST_CASE("Property-based test: VdCorput sequence is deterministic") {
         gen1.reseed(seed);
         gen2.reseed(seed);
 
-        for (size_t i = 0; i < 10; ++i) {
+        for (int i = 0; i < 10; ++i) {
             RC_ASSERT(gen1.pop() == gen2.pop());
         }
     });
@@ -35,7 +35,7 @@ TEST_CASE("Property-based test: VdCorput sequence is deterministic") {
 
 TEST_CASE("Property-based test: VdCorput generates values in [0,1)") {
     ldsgen::VdCorput gen(2);
-    for (size_t i = 0; i < 100; ++i) {
+    for (int i = 0; i < 100; ++i) {
         auto curr = gen.pop();
         CHECK_GE(curr, 0.0);
         CHECK_LT(curr, 1.0);
@@ -44,10 +44,10 @@ TEST_CASE("Property-based test: VdCorput generates values in [0,1)") {
 
 TEST_CASE("Property-based test: Circle points on unit circle") {
     rc::check("All Circle points lie on unit circle", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 101));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 101));
         ldsgen::Circle gen(base);
 
-        for (size_t i = 0; i < 50; ++i) {
+        for (int i = 0; i < 50; ++i) {
             auto point = gen.pop();
             double radius_squared = point[0] * point[0] + point[1] * point[1];
             RC_ASSERT(radius_squared == doctest::Approx(1.0));
@@ -57,11 +57,11 @@ TEST_CASE("Property-based test: Circle points on unit circle") {
 
 TEST_CASE("Property-based test: Disk points inside unit disk") {
     rc::check("All Disk points lie inside unit disk", []() {
-        auto base0 = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto base1 = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base0 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto base1 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
         ldsgen::Disk gen(base0, base1);
 
-        for (size_t i = 0; i < 50; ++i) {
+        for (int i = 0; i < 50; ++i) {
             auto point = gen.pop();
             double radius_squared = point[0] * point[0] + point[1] * point[1];
             RC_ASSERT(radius_squared >= 0.0 && radius_squared <= 1.0);
@@ -71,11 +71,11 @@ TEST_CASE("Property-based test: Disk points inside unit disk") {
 
 TEST_CASE("Property-based test: Sphere points on unit sphere") {
     rc::check("All Sphere points lie on unit sphere", []() {
-        auto base0 = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto base1 = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base0 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto base1 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
         ldsgen::Sphere gen(base0, base1);
 
-        for (size_t i = 0; i < 50; ++i) {
+        for (int i = 0; i < 50; ++i) {
             auto point = gen.pop();
             double radius_squared = point[0] * point[0] + point[1] * point[1] + point[2] * point[2];
             RC_ASSERT(radius_squared == doctest::Approx(1.0));
@@ -85,11 +85,11 @@ TEST_CASE("Property-based test: Sphere points on unit sphere") {
 
 TEST_CASE("Property-based test: Halton points in [0,1]^2") {
     rc::check("All Halton points lie in unit square [0,1]x[0,1]", []() {
-        auto base0 = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto base1 = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base0 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto base1 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
         ldsgen::Halton gen(base0, base1);
 
-        for (size_t i = 0; i < 50; ++i) {
+        for (int i = 0; i < 50; ++i) {
             auto point = gen.pop();
             RC_ASSERT(point[0] >= 0.0 && point[0] < 1.0);
             RC_ASSERT(point[1] >= 0.0 && point[1] < 1.0);
@@ -99,12 +99,12 @@ TEST_CASE("Property-based test: Halton points in [0,1]^2") {
 
 TEST_CASE("Property-based test: Sphere3Hopf points on unit 3-sphere") {
     rc::check("All Sphere3Hopf points lie on unit 3-sphere", []() {
-        auto base0 = static_cast<size_t>(*rc::gen::inRange(2, 31));
-        auto base1 = static_cast<size_t>(*rc::gen::inRange(2, 31));
-        auto base2 = static_cast<size_t>(*rc::gen::inRange(2, 31));
+        auto base0 = static_cast<unsigned long>(*rc::gen::inRange(2, 31));
+        auto base1 = static_cast<unsigned long>(*rc::gen::inRange(2, 31));
+        auto base2 = static_cast<unsigned long>(*rc::gen::inRange(2, 31));
         ldsgen::Sphere3Hopf gen(base0, base1, base2);
 
-        for (size_t i = 0; i < 50; ++i) {
+        for (int i = 0; i < 50; ++i) {
             auto point = gen.pop();
             double radius_squared = point[0] * point[0] + point[1] * point[1] + point[2] * point[2]
                                     + point[3] * point[3];
@@ -115,10 +115,10 @@ TEST_CASE("Property-based test: Sphere3Hopf points on unit 3-sphere") {
 
 TEST_CASE("Property-based test: peek does not advance state") {
     rc::check("peek() returns same value as subsequent pop()", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
         ldsgen::VdCorput gen(base);
 
-        for (size_t i = 0; i < 10; ++i) {
+        for (int i = 0; i < 10; ++i) {
             auto peeked = gen.peek();
             auto popped = gen.pop();
             RC_ASSERT(peeked == popped);
@@ -128,15 +128,15 @@ TEST_CASE("Property-based test: peek does not advance state") {
 
 TEST_CASE("Property-based test: skip advances state correctly") {
     rc::check("skip(n) advances state by n positions", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto skip_count = static_cast<size_t>(*rc::gen::inRange(0, 100));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto skip_count = static_cast<unsigned long>(*rc::gen::inRange(0, 100));
 
         ldsgen::VdCorput gen1(base);
         ldsgen::VdCorput gen2(base);
 
         gen2.skip(skip_count);
 
-        for (size_t i = 0; i < skip_count; ++i) {
+        for (int i = 0; i < skip_count; ++i) {
             gen1.pop();
         }
 
@@ -146,15 +146,15 @@ TEST_CASE("Property-based test: skip advances state correctly") {
 
 TEST_CASE("Property-based test: batch returns correct number of values") {
     rc::check("batch(n) returns exactly n values", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto batch_size = static_cast<size_t>(*rc::gen::inRange(1, 100));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto batch_size = static_cast<unsigned long>(*rc::gen::inRange(1, 100));
 
         ldsgen::VdCorput gen(base);
         auto batch = gen.batch(batch_size);
 
         RC_ASSERT(batch.size() == batch_size);
 
-        for (size_t i = 0; i < batch_size; ++i) {
+        for (int i = 0; i < batch_size; ++i) {
             RC_ASSERT(batch[i] >= 0.0 && batch[i] < 1.0);
         }
     });
@@ -162,8 +162,8 @@ TEST_CASE("Property-based test: batch returns correct number of values") {
 
 TEST_CASE("Property-based test: reseed resets generator state") {
     rc::check("reseed(n) produces deterministic sequence", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto seed = static_cast<size_t>(*rc::gen::inRange(0, 1000));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto seed = static_cast<unsigned long>(*rc::gen::inRange(0, 1000));
 
         ldsgen::VdCorput gen1(base);
         ldsgen::VdCorput gen2(base);
@@ -171,7 +171,7 @@ TEST_CASE("Property-based test: reseed resets generator state") {
         gen1.reseed(seed);
         gen2.reseed(seed);
 
-        for (size_t i = 0; i < 10; ++i) {
+        for (int i = 0; i < 10; ++i) {
             RC_ASSERT(gen1.pop() == gen2.pop());
         }
     });
@@ -179,13 +179,13 @@ TEST_CASE("Property-based test: reseed resets generator state") {
 
 TEST_CASE("Property-based test: iterator produces same values as pop") {
     rc::check("Iterator produces same sequence as pop()", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
 
         ldsgen::VdCorput gen1(base);
         ldsgen::VdCorput gen2(base);
 
         auto it = gen1.begin();
-        for (size_t i = 0; i < 50; ++i, ++it) {
+        for (int i = 0; i < 50; ++i, ++it) {
             RC_ASSERT(*it == gen2.pop());
         }
     });
@@ -193,12 +193,12 @@ TEST_CASE("Property-based test: iterator produces same values as pop") {
 
 TEST_CASE("Property-based test: Circle iterator produces points on unit circle") {
     rc::check("Circle iterator produces points on unit circle", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
 
         ldsgen::Circle gen(base);
         auto it = gen.begin();
 
-        for (size_t i = 0; i < 50; ++i, ++it) {
+        for (int i = 0; i < 50; ++i, ++it) {
             auto point = *it;
             double radius_squared = point[0] * point[0] + point[1] * point[1];
             RC_ASSERT(radius_squared == doctest::Approx(1.0));
@@ -208,13 +208,13 @@ TEST_CASE("Property-based test: Circle iterator produces points on unit circle")
 
 TEST_CASE("Property-based test: Halton iterator produces points in unit square") {
     rc::check("Halton iterator produces points in [0,1]^2", []() {
-        auto base0 = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto base1 = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base0 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto base1 = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
 
         ldsgen::Halton gen(base0, base1);
         auto it = gen.begin();
 
-        for (size_t i = 0; i < 50; ++i, ++it) {
+        for (int i = 0; i < 50; ++i, ++it) {
             auto point = *it;
             RC_ASSERT(point[0] >= 0.0 && point[0] < 1.0);
             RC_ASSERT(point[1] >= 0.0 && point[1] < 1.0);
@@ -224,22 +224,22 @@ TEST_CASE("Property-based test: Halton iterator produces points in unit square")
 
 TEST_CASE("Property-based test: get_index reflects correct position") {
     rc::check("get_index() returns current sequence position", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
 
         ldsgen::VdCorput gen(base);
-        RC_ASSERT(gen.get_index() == static_cast<size_t>(0));
+        RC_ASSERT(gen.get_index() == static_cast<unsigned long>(0));
 
-        for (size_t i = 0; i < 100; ++i) {
+        for (int i = 0; i < 100; ++i) {
             gen.pop();
-            RC_ASSERT(gen.get_index() == i + static_cast<size_t>(1));
+            RC_ASSERT(gen.get_index() == i + static_cast<unsigned long>(1));
         }
     });
 }
 
 TEST_CASE("Property-based test: reseed and get_index consistency") {
     rc::check("reseed(n) sets index to n", []() {
-        auto base = static_cast<size_t>(*rc::gen::inRange(2, 51));
-        auto seed = static_cast<size_t>(*rc::gen::inRange(0, 1000));
+        auto base = static_cast<unsigned long>(*rc::gen::inRange(2, 51));
+        auto seed = static_cast<unsigned long>(*rc::gen::inRange(0, 1000));
 
         ldsgen::VdCorput gen(base);
         gen.reseed(seed);
@@ -247,7 +247,7 @@ TEST_CASE("Property-based test: reseed and get_index consistency") {
         RC_ASSERT(gen.get_index() == seed);
 
         gen.pop();
-        RC_ASSERT(gen.get_index() == seed + static_cast<size_t>(1));
+        RC_ASSERT(gen.get_index() == seed + static_cast<unsigned long>(1));
     });
 }
